+++ ntt/test/ctest/test_ntt_unary_erf_half.cpp
@@ -24,9 +24,9 @@ using namespace ortki;
 TEST(UnaryTestErfFloat, fixed_fixed) {
     // init
     using shape = ntt::fixed_shape<1, 3, 16, 16>;
-    using tensor_type = ntt::tensor<half, shape>;
+    using tensor_type = ntt::tensor<_Float16, shape>;
     std::unique_ptr<tensor_type> ntt_input(new tensor_type);
-    NttTest::init_tensor(*ntt_input, -10.f, 10.f);
+    NttTest::init_tensor(*ntt_input, -static_cast<_Float16>(10.f), static_cast<_Float16>(10.f));
 
     // ntt
     std::unique_ptr<tensor_type> ntt_output1(new tensor_type);
@@ -45,13 +45,13 @@ TEST(UnaryTestErfFloat, fixed_fixed) {
 TEST(UnaryTestErfFloat, fixed_ranked) {
     // init
     using shape1 = ntt::fixed_shape<1, 3, 16, 16>;
-    using tensor_type1 = ntt::tensor<half, shape1>;
+    using tensor_type1 = ntt::tensor<_Float16, shape1>;
     std::unique_ptr<tensor_type1> ntt_input(new tensor_type1);
-    NttTest::init_tensor(*ntt_input, -10.f, 10.f);
+    NttTest::init_tensor(*ntt_input, -static_cast<_Float16>(10.f), static_cast<_Float16>(10.f));
 
     // ntt
     auto shape2 = ntt::make_ranked_shape(1, 3, 16, 16);
-    using tensor_type2 = ntt::tensor<half, ntt::ranked_shape<4>>;
+    using tensor_type2 = ntt::tensor<_Float16, ntt::ranked_shape<4>>;
     std::unique_ptr<tensor_type2> ntt_output1(new tensor_type2(shape2));
     ntt::unary<ntt::ops::erf>(*ntt_input, *ntt_output1);
 
@@ -67,10 +67,10 @@ TEST(UnaryTestErfFloat, fixed_ranked) {
 
 TEST(UnaryTestErfFloat, ranked_ranked) {
     // init
-    using tensor_type = ntt::tensor<half, ntt::ranked_shape<4>>;
+    using tensor_type = ntt::tensor<_Float16, ntt::ranked_shape<4>>;
     auto shape = ntt::make_ranked_shape(1, 3, 16, 16);
     std::unique_ptr<tensor_type> ntt_input(new tensor_type(shape));
-    NttTest::init_tensor(*ntt_input, -10.f, 10.f);
+    NttTest::init_tensor(*ntt_input, -static_cast<_Float16>(10.f), static_cast<_Float16>(10.f));
 
     // ntt
     std::unique_ptr<tensor_type> ntt_output1(new tensor_type(shape));
@@ -89,13 +89,13 @@ TEST(UnaryTestErfFloat, ranked_ranked) {
 TEST(UnaryTestErfFloat, ranked_fixed) {
     // init
     auto shape1 = ntt::make_ranked_shape(1, 3, 16, 16);
-    using tensor_type1 = ntt::tensor<half, ntt::ranked_shape<4>>;
+    using tensor_type1 = ntt::tensor<_Float16, ntt::ranked_shape<4>>;
     std::unique_ptr<tensor_type1> ntt_input(new tensor_type1(shape1));
-    NttTest::init_tensor(*ntt_input, -10.f, 10.f);
+    NttTest::init_tensor(*ntt_input, -static_cast<_Float16>(10.f), static_cast<_Float16>(10.f));
 
     // ntt
     using shape2 = ntt::fixed_shape<1, 3, 16, 16>;
-    using tensor_type2 = ntt::tensor<half, shape2>;
+    using tensor_type2 = ntt::tensor<_Float16, shape2>;
     std::unique_ptr<tensor_type2> ntt_output1(new tensor_type2);
     ntt::unary<ntt::ops::erf>(*ntt_input, *ntt_output1);
 
@@ -133,7 +133,7 @@ template <typename T, size_t vl> void test_vector() {
 #define TEST_VECTOR(T) _TEST_VECTOR(T, 1)
 #endif
 
-TEST(UnaryTestErf, vector) { TEST_VECTOR(half) }
+TEST(UnaryTestErf, vector) { TEST_VECTOR(_Float16) }
 
 template <typename T, size_t vl> void test_vector_ulp(double ulp_threshold) {
     constexpr size_t size = ULP_SIZE;
@@ -176,7 +176,7 @@ template <typename T, size_t vl> void test_vector_ulp(double ulp_threshold) {
 #define TEST_VECTOR_ULP(T, ulp_threshold) _TEST_VECTOR_ULP(T, 1, ulp_threshold)
 #endif
 
-TEST(UnaryTestErfFloat, ulp_error) { TEST_VECTOR_ULP(half, 2.f) }
+TEST(UnaryTestErfFloat, ulp_error) { TEST_VECTOR_ULP(_Float16, 2.f) }
 
 int main(int argc, char *argv[]) {
     ::testing::InitGoogleTest(&argc, argv);
